requirements.txt
dotenv
pydantic-ai
pydantic-ai-slim[groq]
requests
streamlit
